Iteration 2:
The Square class is considered as abstract and special square subclasses are added.
The cash property is added to the Player class.
The association between Player and Piece is now bidirectional.
Other necessary changes are also applied to the Game and Board class.
All the requirements of the 'Applying UML and Patterns' book to the end of Chapter 23 are satisfied.